// services/web-app/app/utils/enums.ts
export enum Period {
  First = '1st',
  Second = '2nd',
  Third = '3rd',
  Fourth = '4th',
  Fifth = '5th',
  Sixth = '6th',
  Seventh = '7th',
  Eighth = '8th',
  Ninth = '9th',
}

export enum Grade {
  Ninth = '9th',
  Tenth = '10th',
  Eleventh = '11th',
  Twelfth = '12th',
}

export enum Setting {
  SignupPasscode = 'signup_passcode',
  Schools = 'schools',
  Teachers = 'teachers',
}